+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - `StackedError` is now special cased to attempt to downcast whenever used as an argument on another `StackedError`, and the two chains are combined to prevent nesting
 - `Default` for `StackedError` uses `Error::empty()` (which does not capture a location and does not allocate) instead of `Error::new()` now
 - Exposed `StackedErrorItem`
+- Improved the macros to accept more expressions
 
 ## [0.7.2] - 2026-08-11
 ### Fixes